Marriage & Parenting Cost Calculator/requirements.txt
customtkinter>=5.2.0
matplotlib>=3.5.0
numpy>=1.21.0
pillow>=8.0.0